requirements.txt
streamlit==1.31.1
google-generativeai==0.4.0
python-dotenv
langchain
PyPDF2
chromadb